% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_plot.R
\name{coexpression_plot}
\alias{coexpression_plot}
\title{Plot Coexpression.}
\usage{
coexpression_plot(data, gene1, gene2)
}
\arguments{
\item{data}{A RNA counts matrix. Columns are sample names. Rows are gene
names.}

\item{gene1}{First gene to plot.}

\item{gene2}{Second gene to plot.}
}
\value{
coexpression plot
}
\description{
A function that plots gene expression of twe genes on same plot.
}
\examples{
cpm_rna_counts <- cpm(rna_counts)
coexpression_plot(cpm_rna_counts, "A1BG", "A2M")


}
