% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\title{Convert A Matrix Of Given RNA Counts To A Matrix Of Counts Per Million.}
\usage{
cpm(matrix)
}
\arguments{
\item{matrix}{a matrix of RNA counts}
}
\value{
Returns a CPM matrix
}
\description{
A function that calculates counts per million(cpm) of each gene for each
sample.
}
\examples{
cpm_rna_counts <- cpm(rna_counts)

rna_counts["A1BG", "Control1"]
# 1
cpm_rna_counts["A1BG", "Control1"]
# 2087.683
 
}
\references{
Koch CM, Chiu SF, Akbarpour M, Bharat A, Ridge KM, Bartom ET, Winter DR.
A Beginner's Guide to Analysis of RNA Sequencing Data.
Am J Respir Cell Mol Biol. 2018. 59(2):145-157.
doi: 10.1165/rcmb.2017-0430TR. PMID: 29624415; PMCID: PMC6096346.
}
