% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Arbitrary RNA Counts matrix.}
\format{
A matrix where columns are the samples and rows are the genes. Each
cell represents RNA counts.
}
\usage{
d
}
\description{
Arbitrary RNA Counts matrix.
}
\keyword{datasets}
