% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_point_plot.R
\name{expression_point_plot}
\alias{expression_point_plot}
\title{Boxplot expression}
\usage{
expression_point_plot(data, gene)
}
\arguments{
\item{data}{RNA counts data.}

\item{gene}{A name of gene.}
}
\value{
plot
}
\description{
A function that boxplots expression of each group of a gene.
}
