% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{Get Fold Change between Control Group And Treatment Group}
\usage{
fold_change(data, gene)
}
\arguments{
\item{data}{A RNA counts matrix.}

\item{gene}{A name of gene to be taken for fold change.}
}
\value{
fold change.
}
\description{
A function that calculates fold change of the RNA counts of gene between
control group and treatment group in log10.
}
\examples{

cpm_d <- cpm(d)
fold_change(cpm_d, "AACS")
#1.459728

}
