% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_all.R
\name{fold_change_all}
\alias{fold_change_all}
\title{Get Fold Change of all genes in the data and sort the genes by fold change
levels.}
\usage{
fold_change_all(data, decreasing = TRUE)
}
\arguments{
\item{data}{A RNA counts matrix.}

\item{decreasing}{Boolean that specifies if the fold is sorted in decreasing
order or not.}
}
\value{
fold change.
}
\description{
A function that calculates fold change of the RNA counts of each gene in
the data and return the sorted list of gene by gene expression level.
}
\examples{
fold_change_result <- fold_change_all(rna_counts)
fold_change_result[fold_change_result$Gene == "AAAS",]
# Gene     Fold

}
