% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_column_plot.R
\name{fold_change_column_plot}
\alias{fold_change_column_plot}
\title{Plot Column Bar Plot Of Fold Changes Of genes.}
\usage{
fold_change_column_plot(fold_change_data)
}
\arguments{
\item{data}{A data frame where first column is 'Gene', second column is
'Fold'. The names of genes will be in the first column and their fold
changes will be in second column.}
}
\value{
column bar plot
}
\description{
A function that plots fold changes of genes in RNA reads counts data in
column bar plot
}
\examples{
dev.off()
fold_change_result <- fold_change_all(rna_counts)
fold_change_column_plot(fold_change_result)

}
