% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_data.R
\name{get_group_data}
\alias{get_group_data}
\title{Get Data Of Either Control Or Treatment Group}
\usage{
get_group_data(data, group)
}
\arguments{
\item{data}{data matrix of RNA counts.}

\item{group}{group of experiment that you want to take out.}
}
\value{
Returns a matrix
}
\description{
A function that filters RNA data of control or treatment group.
}
\examples{
data(library="GeneDiffAnal")
controls <- get_group_data(d, "control")
colnames(controls) # Control1, Control2, Control3, Control4, Control4

}
