% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_t_test.R
\name{group_t_test}
\alias{group_t_test}
\title{Run Welch T Test On Control Group And Treatment Group}
\usage{
group_t_test(data, gene, alternative = "two.sided")
}
\arguments{
\item{data}{A matrix data of RNA counts of control and treatment groups.}

\item{gene}{A name of gene that we want to run t test on RNA expression of.}

\item{alternative}{A character string specifying the alternative hypothesis.
Modify the alternative hypothesis if the control group's expression is less,
the same or greater.}
}
\value{
A list object of t.test result.
}
\description{
A function that runs t.test on RNA expression level
}
\examples{

cpm_rna_counts <- cpm(rna_counts)
t_result <- group_t_test(cpm_rna_counts, "A1BG")
t_result$estimate
# mean of x, mean of y
# 2105.722, 1881.544
# t_result$p.value
# 0.57375

}
\references{
B. L. WELCH. The generalization of student's problem when several
different population variances are involved. Biometrika. 1947. 34(1-2):28-35.
}
