% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_expression.R
\name{heatmap_expression}
\alias{heatmap_expression}
\title{Heatmap of expression of data.}
\usage{
heatmap_expression(data)
}
\arguments{
\item{data}{A RNA counts data matrix where columns are sample names and
row names are gene names. Each cell represents the RNA copy counts of that
column sample and row gene.}
}
\value{
A heatmap of gene expression.
}
\description{
Create a heatmap of expression of genes of samples in RNA level.
}
\examples{
dev.off()
cpm_rna_counts <- cpm(rna_counts)
heatmap_expression(rna_counts)

}
