% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_counts}
\alias{rna_counts}
\title{Arbitrary RNA Counts matrix.}
\format{
A matrix where columns are the samples and rows are the genes. Each
cell represents RNA counts.
}
\usage{
rna_counts
}
\description{
Arbitrary RNA Counts matrix.
}
\examples{
d[1,]
# 1, 2, 1, 1, 1, 1, 2, 1, 1, 2


}
\keyword{datasets}
