% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneDiffAnal.R
\name{runGeneDiffAnal}
\alias{runGeneDiffAnal}
\title{Launch Shiny App for GeneDiffAnal}
\usage{
runGeneDiffAnal()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for GeneDiffAnal.
This app runs interactive user interface app for gene differential analysis.
}
\examples{
\dontrun {
GeneDiffAnal::runGeneDiffAnal()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
