% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_t_test.R
\name{group_t_test}
\alias{group_t_test}
\title{Run Welch T Test On Control Group And Treatment Group}
\usage{
group_t_test(gene, data)
}
\arguments{
\item{gene}{A name of gene that we want to run t test on RNA expression of.}

\item{data}{A matrix data of RNA counts of control and treatment groups.}
}
\value{
A list object of t.test result.
}
\description{
A function that runs t.test on RNA expression level
}
\references{
B. L. WELCH. The generalization of student's problem when several
different population variances are involved. Biometrika. 1947. 34(1-2):28-35.
}
