% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_expression.R
\name{heatmap_expression}
\alias{heatmap_expression}
\title{Heatmap of expression of data.}
\usage{
heatmap_expression(data)
}
\arguments{
\item{data}{Data of RNA counts.}
}
\value{
A heatmap of gene expression.
}
\description{
Create a heatmap of expression of genes of samples in RNA level.
}
\examples{
dev.off()
cpm_d <- cpm(d)
heatmap_expression(cpm_d)

}
