% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_frame.R
\name{matrix_to_frame}
\alias{matrix_to_frame}
\title{Convert RNA Counts Matrix To Data Frame.}
\usage{
matrix_to_frame(matrix)
}
\arguments{
\item{matrix}{}
}
\description{
A function that convert RNA counts matrix into a data frame.
}
\examples{
matrix_to_frame(rna_counts)


}
